% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callingBetaMPT.R
\name{callingBetaMPT}
\alias{callingBetaMPT}
\title{Calling the Sampler (JAGS or OpenBUGS or WinBUGS)}
\usage{
callingBetaMPT(DataFormulas, data, modelfile, numberOfParameters = 0,
  transformedPar = NULL, n.iter = 1e+05, n.burnin = NULL, n.update = 10,
  n.thin = 2, n.chains = 3, sampler = "JAGS", autojags = TRUE, ...)
}
\arguments{
\item{DataFormulas}{Ouptut of mergeBranches()}

\item{data}{data frame with frequencies}

\item{modelfile}{Full path to the model description file. (Made with makeModelDescription())}

\item{numberOfParameters}{How many parameters does the model have}

\item{transformedPar}{vector with names of transformed parameter to be sampled (e.g., parameter differences)}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Burnin period.}

\item{n.update}{Update paramater for JAGS}

\item{n.thin}{Thinning rate.}

\item{n.chains}{number of MCMC chains}

\item{sampler}{The Sampler to be used. Options are JAGS, OpenBUGS or WinBUGS.
Note that you need to install the sampler on your computer.}

\item{autojags}{whether to run JAGS until the MCMC chains converge (see \link{autojags}). Can take a lot of time for large models.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Calling the Sampler (JAGS or OpenBUGS or WinBUGS)
}
\author{
Nina R. Arnold, Denis Arnold, Daniel W. Heck
}

