% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendMPT.R
\name{extendMPT}
\alias{extendMPT}
\title{Extend MCMC Sampling for MPT Model}
\usage{
extendMPT(fittedModel, n.iter = 10000, n.adapt = 1000, n.burnin = 0,
  n.thin = 5, ...)
}
\arguments{
\item{fittedModel}{a fitted \code{\link{traitMPT}} or \code{\link{betaMPT}}}

\item{n.iter}{Number of iterations per chain (including burnin samples).
See \code{\link[runjags]{run.jags}} for details.}

\item{n.adapt}{number of adaption samples to adjust MCMC sampler in JAGS.
The sampler will be more efficient if it is tuned well.}

\item{n.burnin}{Number of samples for burnin (samples will not be stored and removed from n.iter)}

\item{n.thin}{Thinning rate.}

\item{...}{further arguments passed to \link[runjags]{run.jags}}
}
\description{
Extend MCMC Sampling for MPT Model
}
