% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataMPT.R
\name{genBetaMPT}
\alias{genBetaMPT}
\title{Simulate Data for Beta MPT Models}
\usage{
genBetaMPT(N, numItems, eqnfile, mean = NULL, sd = NULL, alpha = NULL,
  beta = NULL)
}
\arguments{
\item{N}{number of participants}

\item{numItems}{number of responses per tree (possibly a named vector with tree labels)}

\item{eqnfile}{The (full path to the) file that specifies the MPT model (standard .eqn syntax)}

\item{mean}{Named vector of true group means of individual MPT parameters. If the vector is not named, the internal order of parameters is used (can be obtained using \code{\link{readEQN}}).}

\item{sd}{named vector of group standard deviations of individual MPT parameters.}

\item{alpha}{Alternative specification of the group-level distribution using the shape parameters of the beta distribution (see \link{dbeta}).}

\item{beta}{see \code{alpha}

Data are generated independently from the JAGS model files used for fitting the Beta-MPT model.}
}
\value{
a list including the generated frequencies (\code{data}) and the true, underlying parameters (\code{parameters})
}
\description{
Generating a data file with known parameter structure using the Beta-MPT. Useful for robustness checks.
}
\examples{
\dontrun{
genDat <- genBetaMPT(N = 100,
                     numItems = c(Target=250, Lure=250),
                     eqnfile = "2htm.eqn",
                     mean = c(Do=.7, Dn=.7, g=.5),
                     sd =   c(Do=.1, Dn=.1, g=.05))
                     }
}

