% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataMPT.R
\name{genMPT}
\alias{genMPT}
\title{Generate MPT Frequencies}
\usage{
genMPT(theta, numItems, eqnfile)
}
\arguments{
\item{theta}{matrix of MPT parameters (rows: individuals; columns: parameters). Parameters are assigned by column names of the matrix. all of the parameters in the model file need to be included.}

\item{numItems}{number of responses per tree (a named vector with tree labels)}

\item{eqnfile}{path to EQN file specifying the MPT model}
}
\description{
Uses a matrix of individual MPT parameters to generate MPT frequencies.
}
\seealso{
\code{\link{genTraitMPT}} and \code{\link{genBetaMPT}} to generate data for latent normal/beta hierarchical distributions.
}

