% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CollectionOfInputProcessingFunctions.R
\name{getParameter}
\alias{getParameter}
\title{Extract all parameter from the formulas of a given model}
\usage{
getParameter(TreeData)
}
\arguments{
\item{TreeData}{Data returned from readMultiTree()}
}
\description{
Extract all parameter from the formulas of a given model
}
\author{
Nina R. Arnold, Denis Arnold
}

