% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeModelDescription.R
\name{makeModelDescription}
\alias{makeModelDescription}
\title{Make the model description file needed by WinBUGS.}
\usage{
makeModelDescription(filename, DataFormulas, nthetas, alpha = "dunif(1,5000)",
  beta = "dunif(1,5000)", sampler = "JAGS", parString = "")
}
\arguments{
\item{filename}{The name that the model description file should get}

\item{DataFormulas}{Output of mergeBranches()}

\item{nthetas}{The number of thetas}

\item{alpha}{Value for alpha}

\item{beta}{Value for betha}

\item{sampler}{Which sampler should be used? Default is "JAGS", further options are "OpenBUGS" and "WinBugs"}

\item{parString}{additional lines added to sample transformed parameters (e.g., parameter differences for testing)}
}
\description{
Make the model description file needed by WinBUGS.
}
\author{
Nina R. Arnold, Denis Arnold, Daniel Heck
}

