% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CollectionOfInputProcessingFunctions.R
\name{mergeBranches}
\alias{mergeBranches}
\title{Unique branches for each Tree with the same answer and sum the corresponing formulas}
\usage{
mergeBranches(TreeData)
}
\arguments{
\item{TreeData}{Data returned from readMultiTree()}
}
\description{
Unique branches for each Tree with the same answer and sum the corresponing formulas
}
\author{
Nina R. Arnold, Denis Arnold
}

