% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpt2BetaMPT.R
\name{mpt2BetaMPT}
\alias{mpt2BetaMPT}
\title{Userinterface to use BUGS}
\usage{
mpt2BetaMPT(eqnfile, subjdata, restrictions = NULL, modelfilename = NULL,
  alpha = "dunif(1,5000)", beta = "dunif(1,5000)",
  parameters = list("theta", "alph", "bet", "mnb", "varp"), parestfile,
  n.iter = 1e+05, n.burnin = 50000, n.thin = 2, sampler = "JAGS",
  autojags = TRUE, ...)
}
\arguments{
\item{eqnfile}{The (full path to the) file that specifies the Multitree.}

\item{subjdata}{The (full path to the) cvs file with the data.}

\item{restrictions}{The (full path to the) file that specifies which parameters should be constants and which should be equal.}

\item{modelfilename}{Name that the modelfile that is made by the function to work with WinBUGS should get.
Default is to write this information to the tempdir as required by CRAN standards.}

\item{alpha}{Model parameter.}

\item{beta}{Model parameter.}

\item{parameters}{List of parameters that WinBUGS should return.}

\item{parestfile}{Name of the file to which the results should be stored.}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Burnin period.}

\item{n.thin}{Thinning rate.}

\item{sampler}{Which sampler should be used? Default is "JAGS", further options are "OpenBUGS" and "WinBugs"}

\item{autojags}{whether to run JAGS until the MCMC chains converge (see \link{autojags}). Can take a lot of time for large models.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Userinterface to use BUGS
}
\author{
Nina R. Arnold, Denis Arnold
}

