% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpt2BetaMPT_plots.R
\name{plot.betaMPT}
\alias{plot.betaMPT}
\title{Plot Beta-MPT Parameter Estimates}
\usage{
\method{plot}{betaMPT}(x, includeIndividual = TRUE, ...)
}
\arguments{
\item{x}{a fitted Beta-MPT model (see \code{\link{mpt2BetaMPT}})}

\item{includeIndividual}{whether to plot individual estimates}

\item{...}{further arguments passed to the standard \code{\link{plot}} function}
}
\description{
Plot Beta-MPT Parameter Estimates
}
\author{
Daniel Heck
}

