% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMPT_plots.R
\name{plot.traitMPT}
\alias{plot.traitMPT}
\title{Plots for latent-trait MPT models (see \code{\link{traitMPT}}).}
\usage{
\method{plot}{traitMPT}(x, includeIndividual = TRUE, ...)
}
\arguments{
\item{x}{a fitted Beta-MPT model (see \code{\link{betaMPT}})}

\item{includeIndividual}{whether to plot individual estimates}

\item{...}{further arguments passed to the standard \code{\link{plot}} function}
}
\description{
Plots for latent-trait MPT models (see \code{\link{traitMPT}}).
}
\author{
Daniel Heck
}

