% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plotFit}
\alias{plotFit}
\title{Plot Goodness of Fit}
\usage{
plotFit(fittedModel, ...)
}
\arguments{
\item{fittedModel}{fitted latent-trait or beta MPT model (\code{\link{traitMPT}})}

\item{...}{arguments passed to \code{\link{boxplot}}}
}
\description{
Plots observed mean frequencies against sampled mean frequencies.
}

