% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriorPost.R
\name{plotPriorPost}
\alias{plotPriorPost}
\title{Plot Prior vs. Posterior Distribution}
\usage{
plotPriorPost(fittedModel, M = 2e+05, ci = 0.95)
}
\arguments{
\item{fittedModel}{fitted latent-trait or beta MPT model (\code{\link{traitMPT}}, \code{\link{betaMPT}})}

\item{M}{number of random samples to approximate prior distributions}

\item{ci}{credibility interval indicated by vertical red lines}
}
\description{
Allows to judge how much the data informed the parameter posterior distributions compared to the prior.
}
\details{
Prior distributions are shown as blue, dashed lines, whereas posterior distributions are shown as solid, black lines.
}

