% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorPredictive.R
\name{priorPredictive}
\alias{priorPredictive}
\title{Prior Predictive Samples}
\usage{
priorPredictive(prior, eqnfile, restrictions, numItems, N = 1, M = 100,
  nCPU = 4)
}
\arguments{
\item{prior}{a named list defining the priors. For the \link{traitMPT}, the default is \code{list(mu = "dnorm(0,1)", xi="dunif(0,10)", V=diag(S), df=S+1)}, where S is the number of free parameters. For the \link{betaMPT}, the default is \code{list(alpha ="dgamma(1,.1)", beta = "dgamma(1,.1)")}}

\item{eqnfile}{The (full path to the) file that specifies the MPT model (standard .eqn syntax). Note that category labels must start with a letter (different to multiTree) and match the column names of \code{data}. Alternatively, the EQN-equations can be provided within R as a character value (cf. \code{\link{readEQN}})}

\item{restrictions}{Specifies which parameters should be (a) constant (e.g., \code{"a=b=.5"}) or (b) constrained to be identical (e.g., \code{"Do=Dn"}) or (c) treated as fixed effects (i.e., identical for all participants; \code{"a=b=FE"}). Either given as the path to a text file with restrictions per row or as a list of restrictions, e.g., \code{list("D1=D2","g=0.5")}}

\item{numItems}{vector with the number of items per MPT tree (either named or assigned to alphabetically ordered tree labels)}

\item{N}{number of participants}

\item{M}{number of prior predictive samples (i.e., complete data sets).}

\item{nCPU}{number of CPUs used for parallel sampling. For large models and many participants, this requires a lot of memory.}
}
\value{
a list of \code{M} matrices with individual frequencies (rows=participants, columns=MPT categories). For \code{M=1}, a single matrix is returned.
}
\description{
Sample full data sets (i.e., individual response frequencies) based on prior distribution for group-level parameters.
}
\examples{
eqnfile <- system.file("MPTmodels/2htm.eqn",
                       package="TreeBUGS")
prior <- list(alpha="dgamma(1,.1)",
              beta="dgamma(1,.1)")
priorPredictive(prior, eqnfile,
                restrictions=list("g=.5","Do=Dn"),
                numItems=c(50,50), N=25, M=1, nCPU=1)

}

