% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CollectionOfInputProcessingFunctions.R
\name{readMultiTree}
\alias{readMultiTree}
\title{Read a multiTree file}
\usage{
readMultiTree(file)
}
\arguments{
\item{file}{The (full path to the) file that specifies the Multitree}
}
\description{
Read a multiTree file
}
\author{
Nina R. Arnold, Denis Arnold
}

