% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CollectionOfInputProcessingFunctions.R
\name{readSubjectData}
\alias{readSubjectData}
\title{Read the subject data from file}
\usage{
readSubjectData(data, answers)
}
\arguments{
\item{data}{the data as data.frame}

\item{answers}{is the unique of the $Answers from TreeData which is $Answers after mergeBrachnes()}
}
\description{
Read the subject data from file
}
\author{
Nina R. Arnold, Denis Arnold
}

