% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMCMC.R
\name{summarizeMCMC}
\alias{summarizeMCMC}
\title{MCMC Summary}
\usage{
summarizeMCMC(mcmc, batchSize = 100)
}
\arguments{
\item{mcmc}{a \code{\link[coda]{mcmc.list}} object}

\item{batchSize}{size of batches of parameters used to reduce memory load when computing (univariate) Rhat statistics}
}
\description{
TreeBUGS-specific summary for \code{mcmc.list}-objects.
}

