% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CollectionOfInputProcessingFunctions.R
\name{thetaHandling}
\alias{thetaHandling}
\title{Set constants, replace parameters for thetas}
\usage{
thetaHandling(MergedTree, file)
}
\arguments{
\item{MergedTree}{is the output of mergeBranches}

\item{file}{The (full path to the) file that specifies which parameters should be constants and which should be equal}
}
\description{
Set constants, replace parameters for thetas
}
\author{
Nina R. Arnold, Denis Arnold
}

