% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMPT.R
\name{betaMPT}
\alias{betaMPT}
\title{Fit a Hierarchical Beta-MPT Model}
\usage{
betaMPT(eqnfile, data, restrictions, covData, covStructure,
  transformedParameters, alpha = "dunif(.01,5000)",
  beta = "dunif(.01,5000)", n.iter = 50000, n.burnin = 5000, n.thin = 5,
  n.chains = 3, modelfilename, parEstFile, autojags = FALSE, ...)
}
\arguments{
\item{eqnfile}{The (full path to the) file that specifies the MPT model (standard .eqn syntax)}

\item{data}{The (full path to the) csv file with the data (semicolon separated; category labels in first row). Alternatively: a data frame or matrix (rows=individuals, columns = individual category frequencies, category labels as column names)}

\item{restrictions}{Optional: Either the (full path to the) file that specifies which parameters should be constants and which should be equal; alternatively: a list of restrictions, e.g., \code{list("D1=D2","g=0.5")}}

\item{covData}{The path to the csv data file for the individual values on the covariates (semicolon-separated; rows=individuals in the same order as \code{data}, covariate labels in first row). Alternatively: a data frame or matrix (rows=individuals, columns = individual values on covariates, covariate labels as column names)}

\item{covStructure}{Optional: Either the (full path to the) file that specifies the assigment of covariates to MPT parameters (that is, each row assigns one covariate to one or more parameters separated by a semicolon, e.g., \code{age ; Do Dn}). Can also be provided as a list, e.g., \code{list("age ; Do Dn", "extraversion ; g"}). Default: All combinations included (could be unstable).}

\item{transformedParameters}{list with parameter transformations that should be computed based on the posterior samples (e.g., for testing parameter differences: \code{list("diffD=Do-Dn")})}

\item{alpha}{Hyperprior of for the alpha and beta parameters (default: uniform prior on the interval [1,5000]).}

\item{beta}{Second hyperprior, see \code{alpha}}

\item{n.iter}{Number of iterations (including burnin samples).}

\item{n.burnin}{Burnin period.}

\item{n.thin}{Thinning rate.}

\item{n.chains}{number of MCMC chains}

\item{modelfilename}{Name that the modelfile that is made by the function to work with JAGS should get.
Default is to write this information to the tempdir as required by CRAN standards.}

\item{parEstFile}{Name of the file to with the estimates should be stored (e.g., "parEstFile.txt")}

\item{autojags}{whether to run JAGS until the MCMC chains converge (see \link{autojags}).  Use \code{n.update=3} as an additional argument to control how often JAGS is rerun. Can take a lot of time for large models.}

\item{...}{Arguments to be passed to the sampling function (default: \code{\link{jags.parallel}}.}
}
\value{
a list of the class \code{betaMPT} with the objects:
\itemize{
 \item \code{summary}: MPT tailored summary. Use \code{summary(fittedModel)}
 \item \code{mptInfo}: info about MPT model (eqn and data file etc.)
 \item \code{mcmc}: the object returned from the MCMC sampler. Standard: An \code{\link{jags.parallel}} object. Note that the sample can be transformed into an \code{mcmc.list} for analysis with the \code{coda} package by \code{as.mcmc.list(fittedModel$mcmc$BUGSoutput)}
}
}
\description{
Fits a Beta-MPT model (Smith & Batchelder, 2010) based on a standard MPT model file (.eqn) and individual data table (.csv).
}
\details{
Note that, in the Beta-MPT model, correlations of individual MPT parameters with covariates are sampled. Hence, the covariates do not affect the estimation of the actual Beta-MPT parameters. Therefore, the correlation of covariates with the individual MPT parameters can equivalently be performed after fitting the model using the sampled posterior parameter values stored in \code{betaMPT$mcmc}
}
\author{
Nina R. Arnold, Denis Arnold, Daniel Heck
}
\references{
Smith, J. B., & Batchelder, W. H. (2010). Beta-MPT: Multinomial processing tree models for addressing individual differences. Journal of Mathematical Psychology, 54, 167-183.
}

