% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callingBetaMPT.R
\name{callingBetaMPT}
\alias{callingBetaMPT}
\title{Calling the Sampler (JAGS or OpenBUGS or WinBUGS)}
\usage{
callingBetaMPT(DataFormulas, datafile, modelfile, numberOfParameters = 0,
  parameters = list("theta", "alph", "bet", "mnb", "varp"),
  parestfile = "ResultsBetaMPT.txt", n.iter = 1e+05, n.burnin = NULL,
  n.update = 10, n.thin = 2, sampler = "JAGS", printresults = FALSE,
  savetable = FALSE, ...)
}
\arguments{
\item{DataFormulas}{Ouptut of mergeBranches()}

\item{datafile}{Full path to cvs file with the data.}

\item{modelfile}{Full path to the model description file. (Made with makeModelDescription())}

\item{numberOfParameters}{How many parameters does the model have}

\item{parameters}{List with parameters that should be returned from the sampler.}

\item{parestfile}{Name of the file to with the estimates should be stored.}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Burnin period.}

\item{n.update}{Update paramater for JAGS}

\item{n.thin}{Thinning rate.}

\item{sampler}{The Sampler to be used. Options are JAGS, OpenBUGS or WinBUGS.
       Note that you need to install the sampler on your computer.}

\item{printresults}{Whether a summary of the results should be printed.}

\item{savetable}{Whether the results should be saved to the hard drive.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Calling the Sampler (JAGS or OpenBUGS or WinBUGS)
}
\author{
Nina R. Arnold, Denis Arnold
}

