% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMCMC.R
\name{summarizeMCMC}
\alias{summarizeMCMC}
\title{MCMC Summary}
\usage{
summarizeMCMC(mcmc, batchSize = 50, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{mcmc}{a \code{\link[coda]{mcmc.list}} object}

\item{batchSize}{size of batches of parameters used to reduce memory load when
computing posterior summary statistics (including Rhat and effective sample size).}

\item{probs}{quantile probabilities used to compute credibility intervals}
}
\description{
TreeBUGS-specific MCMC summary for \code{mcmc.list}-objects.
}
