% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitMPT.R
\name{traitMPT}
\alias{traitMPT}
\title{Fit a Hierarchical latent-trait MPT Model}
\usage{
traitMPT(eqnfile, data, restrictions = NULL, transformedParameters = NULL,
  mu = "dnorm(0,1)", xi = "dunif(0,100)", V = NULL, df = NULL,
  n.iter = 1e+05, n.burnin = 50000, n.thin = 2, n.chains = 3,
  modelfilename = NULL, parEstFile = NULL, sampler = "JAGS",
  autojags = TRUE, ...)
}
\arguments{
\item{eqnfile}{The (full path to the) file that specifies the MPT model (standard .eqn syntax)}

\item{data}{The (full path to the) cvs file with the data (comma separated; category labels in first row)}

\item{restrictions}{Optional: The (full path to the) file that specifies which parameters should be constants and which should be equal. Alternatively: a list of restrictions, e.g., \code{list("D1=D2","g=0.5")}}

\item{transformedParameters}{list with parameter transformations that should be computed based on the posterior samples (e.g., for testing parameter differences: \code{list("diffD=Do-Dn")})}

\item{mu}{hyperprior for group means of probit-transformed parameters. Default is a standard normal distribution.}

\item{xi}{hyperprior for scaling parameters of the group-level parameter variances. Default is a uniform distribution on the interval [0,100]}

\item{V}{S x S matrix used as a hyperprior for the inverse-Wishart hyperprior parameters with as many rows and columns as there are core MPT parameters. Default is a diagonal matrix.}

\item{df}{degrees of freedom for the inverse-Wishart hyperprior for the individual parameters. Minimum is S+1, where S gives the number of core MPT parameters.}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Burnin period.}

\item{n.thin}{Thinning rate.}

\item{n.chains}{number of MCMC chains}

\item{modelfilename}{Name that the modelfile that is made by the function to work with WinBUGS should get.
Default is to write this information to the tempdir as required by CRAN standards.}

\item{parEstFile}{Name of the file to with the estimates should be stored (e.g., "parEstFile.txt")}

\item{sampler}{Which sampler should be used? Default is "JAGS". Further options are "OpenBUGS" and "WinBugs" (without MPT specific summary and plotting functions)}

\item{autojags}{whether to run JAGS until the MCMC chains converge (see \link{autojags}). Can take a lot of time for large models.}

\item{...}{Arguments to be passed to the sampling function (default: \code{\link{jags.parallel}}.}
}
\value{
a list of the class \code{traitMPT} with the objects:
\itemize{
 \item \code{summary}: MPT tailored summary. Use \code{summary(fittedModel)}
 \item \code{mptInfo}: info about MPT model (eqn and data file etc.)
 \item \code{mcmc}: the object returned from the MCMC sampler. Standard: An \code{\link{jags.parallel}} object. Note that the sample can be transformed into an \code{mcmc.list} for analysis with the \code{coda} package by \code{as.mcmc.list(fittedModel$mcmc$BUGSoutput)}
 \item \code{sampler}: the type of sampler used (standard: \code{"JAGS"})
}
}
\description{
Fits a latent-trait MPT model (Klauer, 2010) based on a standard MPT model file (.eqn) and individual data table (.csv).
}
\author{
Daniel Heck, Denis Arnold, Nina R. Arnold
}
\references{
Klauer, K. C. (2010). Hierarchical multinomial processing tree models: A latent-trait approach. Psychometrika, 75, 70-98.
}

