% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitMPT.R
\name{traitMPT}
\alias{traitMPT}
\title{Fit a Hierarchical latent-trait MPT Model}
\usage{
traitMPT(eqnfile, data, restrictions, covData, covStructure, covType,
  transformedParameters, mu = "dnorm(0,1)", xi = "dunif(0,100)", V, df,
  n.iter = 50000, n.burnin = 5000, n.thin = 5, n.chains = 3,
  modelfilename, parEstFile, sampler = "JAGS", autojags = FALSE, ...)
}
\arguments{
\item{eqnfile}{The (full path to the) file that specifies the MPT model (standard .eqn syntax)}

\item{data}{The (full path to the) csv file with the data (semicolon separated; category labels in first row). Alternatively: a data frame or matrix (rows=individuals, columns = individual category frequencies, category labels as column names)}

\item{restrictions}{Optional: Either the (full path to the) file that specifies which parameters should be constants and which should be equal; alternatively: a list of restrictions, e.g., \code{list("D1=D2","g=0.5")}}

\item{covData}{The path to the csv data file for the individual values on the covariates (semicolon-separated; rows=individuals in the same order as \code{data}, covariate labels in first row). Alternatively: a data frame or matrix (rows=individuals, columns = individual values on covariates, covariate labels as column names)}

\item{covStructure}{Optional: Either the (full path to the) file that specifies the assigment of covariates to MPT parameters (that is, each row assigns one covariate to one or more parameters separated by a semicolon, e.g., \code{age ; Do Dn}). Can also be provided as a list, e.g., \code{list("age ; Do Dn", "extraversion ; g"}). Default: All combinations included (could be unstable).}

\item{covType}{a character vector specifying the type of covariates in each column of \code{covData}: \code{"c"} = continuous covariate (which is the default for continuous variables); \code{"f"} = fixed effect (default for character/factor variables); \code{"r"} = random effect.}

\item{transformedParameters}{list with parameter transformations that should be computed based on the posterior samples (e.g., for testing parameter differences: \code{list("diffD=Do-Dn")})}

\item{mu}{hyperprior for group means of probit-transformed parameters. Default is a standard normal distribution.}

\item{xi}{hyperprior for scaling parameters of the group-level parameter variances. Default is a uniform distribution on the interval [0,100]}

\item{V}{S x S matrix used as a hyperprior for the inverse-Wishart hyperprior parameters with as many rows and columns as there are core MPT parameters. Default is a diagonal matrix.}

\item{df}{degrees of freedom for the inverse-Wishart hyperprior for the individual parameters. Minimum is S+1, where S gives the number of core MPT parameters.}

\item{n.iter}{Number of iterations (including burnin samples).}

\item{n.burnin}{Burnin period.}

\item{n.thin}{Thinning rate.}

\item{n.chains}{number of MCMC chains}

\item{modelfilename}{Name that the modelfile that is made by the function to work with WinBUGS should get.
Default is to write this information to the tempdir as required by CRAN standards.}

\item{parEstFile}{Name of the file to with the estimates should be stored (e.g., "parEstFile.txt")}

\item{sampler}{Which sampler should be used? Default is "JAGS". Further options are "OpenBUGS" and "WinBugs" (without MPT specific summary and plotting functions)}

\item{autojags}{whether to run JAGS until the MCMC chains converge (see \link{autojags}). Can take a lot of time for large models.}

\item{...}{Arguments to be passed to the sampling function (default: \code{\link{jags.parallel}}.}
}
\value{
a list of the class \code{traitMPT} with the objects:
\itemize{
 \item \code{summary}: MPT tailored summary. Use \code{summary(fittedModel)}
 \item \code{mptInfo}: info about MPT model (eqn and data file etc.)
 \item \code{mcmc}: the object returned from the MCMC sampler. Standard: An \code{\link{jags.parallel}} object. Note that the sample can be transformed into an \code{mcmc.list} for analysis with the \code{coda} package by \code{as.mcmc.list(fittedModel$mcmc$BUGSoutput)}
 \item \code{sampler}: the type of sampler used (standard: \code{"JAGS"})
}
}
\description{
Fits a latent-trait MPT model (Klauer, 2010) based on a standard MPT model file (.eqn) and individual data table (.csv).
}
\author{
Daniel Heck, Denis Arnold, Nina R. Arnold
}
\references{
Klauer, K. C. (2010). Hierarchical multinomial processing tree models: A latent-trait approach. Psychometrika, 75, 70-98.

Matzke, D., Dolan, C. V., Batchelder, W. H., & Wagenmakers, E.-J. (2015). Bayesian estimation of multinomial processing tree models with heterogeneity in participants and items. Psychometrika, 80, 205-235.
}

